refactor: delete inviteController
